FEAT: implement AFS (#456)
closes #455 
closes #345

---------

+++ source/types.hpp
@@ -15,7 +15,7 @@ using PacketNum = std::uint32_t;
 using TTL = std::uint32_t;
 
 // Describes a type used by packet's bitset to store flags
-using PacketFlagsBase = std::uint32_t;
+using PacketFlagsBase = std::uint64_t;
 
 template <typename K, typename V>
 using MapWeakPtr =
